grid: fast version of nbn and ebe (gridred), int to size_t
git-svn-id: svn+ssh://svn.code.sf.net/p/nuto/code/trunk@604 b5606e62-d4ae-482b-8a79-2ac8f2527b1c

+++ src/nuto/mechanics/structures/grid/StructureGrid.h
@@ -49,7 +49,7 @@ public:
 #endif // ENABLE_SERIALIZATION
 
     //! @brief import routine for basic grid data without StructureGrid data space
-    void ImportFromVtkASCIIFileHeader(const char* rFileName,int *rGridDimension,double *rVoxelSpacing,double *rGridOrigin, int &rNumVoxel);
+    void ImportFromVtkASCIIFileHeader(const char* rFileName,size_t *rGridDimension,double *rVoxelSpacing,double *rGridOrigin, size_t &rNumVoxel);
 
     //! @brief import routine for basic grid data with StructureGrid data space
     void ImportFromVtkASCIIFileHeader(const char* rFileName);
